added another surface test and moved helper functions into surface_test_helpers.py
